+++ src/Adapter/Product/Update/ProductCustomizationFieldUpdater.php
@@ -26,9 +26,10 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Product;
+namespace PrestaShop\PrestaShop\Adapter\Product\Update;
 
 use CustomizationField;
+use PrestaShop\PrestaShop\Adapter\Product\CustomizationFieldDeleter;
 use PrestaShop\PrestaShop\Adapter\Product\Repository\CustomizationFieldRepository;
 use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductRepository;
 use PrestaShop\PrestaShop\Core\Domain\Product\Customization\ValueObject\CustomizationFieldId;
